refactor(entity): Car, Cars, RaceResult 를 entity 폴더로 이동
