+++ src/main/java/racingcar/domain/entity/Cars.java
@@ -1,4 +1,4 @@
-package racingcar.domain;
+package racingcar.domain.entity;
 
 import java.util.Arrays;
 import java.util.List;
