+++ src/main/java/racingcar/domain/entity/RaceResult.java
@@ -1,4 +1,4 @@
-package racingcar.domain;
+package racingcar.domain.entity;
 
 import java.util.List;
