+++ src/test/java/racingcar/domain/entity/RaceResultTest.java
@@ -1,4 +1,4 @@
-package racingcar.domain;
+package racingcar.domain.entity;
 
 import static org.assertj.core.api.Assertions.assertThat;
