+++ src/test/java/racingcar/service/RaceServiceTest.java
@@ -5,7 +5,7 @@ import static org.assertj.core.api.Assertions.assertThat;
 import java.util.List;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
-import racingcar.domain.Cars;
+import racingcar.domain.entity.Cars;
 import racingcar.domain.strategy.AlwaysMoveStrategy;
 import racingcar.domain.strategy.MoveStrategy;
 import racingcar.dto.RoundResultDto;
