+++ src/test/java/racingcar/util/FormatUtilTest.java
@@ -7,8 +7,8 @@ import java.util.List;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.ValueSource;
-import racingcar.domain.Car;
-import racingcar.domain.Cars;
+import racingcar.domain.entity.Car;
+import racingcar.domain.entity.Cars;
 import racingcar.domain.dto.CarSnapShot;
 import racingcar.domain.strategy.AlwaysMoveStrategy;
 import racingcar.dto.WinnerDto;
